First steps of refactor
